requirements.txt
playwright
psycopg2-binary
phonenumbers
python-dotenv
requests